+++ CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.1.0-alpha.2](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v5.1.0-alpha.1...v5.1.0-alpha.2) (2026-04-01)
+
+
+### Features
+
+* add password visibility toggle and loading state to login form ([e69c4fb](https://github.com/immobiliare/backstage-plugin-ldap-auth/commit/e69c4fb5311f5858d7c79ce76a4353fdb0a395cc))
+
+
+
+
+
+# [5.1.0-alpha.1](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v5.1.0-alpha.0...v5.1.0-alpha.1) (2026-04-01)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+# [5.1.0-alpha.0](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v5.0.0...v5.1.0-alpha.0) (2026-04-01)
+
+
+### Bug Fixes
+
+* **provider:** prevent token refresh from invalidating concurrent sessions ([aca0408](https://github.com/immobiliare/backstage-plugin-ldap-auth/commit/aca040859ede323a861ea61f8dd3a83f6bd89636))
+
+
+### Features
+
+* updated dependencies ([9b8863b](https://github.com/immobiliare/backstage-plugin-ldap-auth/commit/9b8863ba2fdbdfb7d35414b6fb2ae130c7b98dd0))
+
+
+
+
+
 # [5.0.0](https://github.com/immobiliare/backstage-plugin-ldap-auth/compare/v5.0.0-alpha.6...v5.0.0) (2026-03-25)
 
 **Note:** Version bump only for package root